Rag_csv/requirements.txt
langchain_openai 
python-dotenv
streamlit
langchain_community
python-dotenv
streamlit
faiss-cpu
